Author: R Phillips <r.phillips@uq.edu.au> Bug 2924: RADIUS helper compile issues

+++ helpers/basic_auth/squid_radius_auth/squid_rad_auth.c
@@ -572,7 +572,7 @@ main(int argc, char **argv)
 
         /* Parse out the username and password */
         ptr = authstring;
-        while (isspace(*ptr))
+        while (isspace((int)*ptr))
             ptr++;
         if ((end = strchr(ptr, ' ')) == NULL) {
             printf("ERR\n");	/* No password */
@@ -581,7 +581,7 @@ main(int argc, char **argv)
         *end = '\0';
         urldecode(username, ptr, MAXPWNAM);
         ptr = end + 1;
-        while (isspace(*ptr))
+        while (isspace((int)*ptr))
             ptr++;
         urldecode(passwd, ptr, MAXPASS);
